test: share checkpoint boundary fixture
